ase project group-16/requirements.txt
asgiref==3.2.3
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
decorator==4.4.1
defusedxml==0.6.0
django==3.0.2
django-crispy-forms==1.8.1
django-filter==2.2.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
entrypoints==0.3
Flask==1.1.1
Flask-WTF==0.14.2
graphviz==0.13.2
idna==2.9
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.2
locationiq==0.0.2
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
phonenumbers==8.11.5
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
Pygments==2.5.2
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.23.0
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
virtualenv==16.7.7
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
WTForms==2.2.1
xlrd==1.2.0
